+++ lib/dollar.rb
@@ -1,18 +1,8 @@
-class Dollar
+require 'money'
 
-  def initialize(amount)
-    @amount = amount
-  end
+class Dollar < Money
 
   def times(multiplier)
     self.class.new(@amount * multiplier)
   end
-
-  def ==(other)
-    self.amount == other.amount
-  end
-
-  protected
-
-    attr_reader :amount
 end
